requirements.txt
flask
tensorflow
numpy
Pillow
gunicorn
flask_cors
flask-cors

